- Distributed Transactions Example using ChainedTransactionManager - InterestService with custom integrity check (3 interest as maximum) - InterestControllerIntegrationTest assuring transaction gets rollbecked for product 4

+++ src/main/java/org/nh/rest/persistence/relational/ds02/ProductRepository.java
@@ -1,10 +1,9 @@
 package org.nh.rest.persistence.relational.ds02;
 
 import org.nh.rest.model.ds02.Product;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     Product findByName(String name);
